+++ package.json
@@ -25,6 +25,7 @@
   },
   "dependencies": {
     "@apollo/client": "^4.1.6",
+    "@apollo/client-integration-nextjs": "^0.14.4",
     "@emotion/cache": "^11.14.0",
     "@emotion/react": "^11.14.0",
     "@emotion/server": "^11.11.0",
