+++ src/app/root/[domain]/[lang]/(with-layout)/layout.tsx
@@ -1,0 +1,11 @@
+import type { ReactNode } from 'react';
+
+import Layout from '@/components/Layout';
+
+type Props = {
+  children: ReactNode;
+};
+
+export default function WithLayoutLayout({ children }: Props) {
+  return <Layout>{children}</Layout>;
+}
