+++ src/app/root/[domain]/[lang]/(with-layout)/page.tsx
@@ -1,0 +1,7 @@
+'use client';
+
+import Page from '@/components/pages/Page';
+
+export default function HomePage() {
+  return <Page path="/" />;
+}
