+++ src/app/root/[domain]/[lang]/loading.tsx
@@ -1,0 +1,7 @@
+'use client';
+
+import ContentLoader from '@/components/common/ContentLoader';
+
+export default function Loading() {
+  return <ContentLoader fullPage />;
+}
