--- src/pages/404.tsx
@@ -1,5 +1,0 @@
-import Error from './_error';
-
-export default function NotFoundError() {
-  return <Error statusCode={404} />;
-}
