--- src/pages/500.tsx
@@ -1,5 +1,0 @@
-import Error from './_error';
-
-export default function InternalServerError() {
-  return <Error statusCode={500} />;
-}
